Move accompanying persons fixture
